core_requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile core_requirements.in -o core_requirements.txt
anyio==4.10.0
    # via httpx
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
colorama==0.4.6
    # via loguru
google-api-core==2.25.1
    # via google-api-python-client
google-api-python-client==2.181.0
    # via -r core_requirements.in
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r core_requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.2
    # via -r core_requirements.in
googleapis-common-protos==1.70.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.30.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via -r core_requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
loguru==0.7.3
    # via -r core_requirements.in
oauthlib==3.3.1
    # via requests-oauthlib
proto-plus==1.26.1
    # via google-api-core
protobuf==6.32.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pyparsing==3.2.3
    # via httplib2
requests==2.32.5
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
win32-setctime==1.2.0
    # via loguru
